requirements.txt
bs4
requests
python-dotenv
git+https://github.com/nficano/pytube.git
